+++ src/kineverse/utils.py
@@ -2,6 +2,7 @@ import os
 
 import copy  as copy_module
 import math
+import matplotlib.colors as matcolors
 import numpy as np
 import kineverse.gradients.gradient_math as gm
 
@@ -159,3 +160,170 @@ def generate_transition_function(sym_dt, symbols, overrides=None):
     function = gm.speed_up(temp_f, f_params)
 
     return ordered_symbols, function, f_params
+
+
+BASE_COLORS = [(35, 95, 100), (355, 95, 91), (261, 100, 100), (194, 95, 91), (123, 100, 100)]#, 
+               #(15, 95, 100), (293, 95, 91), (244, 100, 100), (156, 95, 91), (76, 100, 100)]
+BASE_COLORS = [np.array(b) * np.array([1.0/360, 1e-2, 1e-2]) for b in BASE_COLORS]
+BASE_COLORS = [[b * (1.0, 1.0, f) for f in np.linspace(1.0, 0.5, 3)] for b in BASE_COLORS]
+
+def hsv_to_rgb(h, s, v):
+    return matcolors.hsv_to_rgb((h, s, v))
+
+def rgb_to_hex(r, g, b):
+    return '#{:02X}{:02X}{:02X}'.format(int(r * 255), int(g * 255), int(b * 255))
+
+def print_return(f):
+    def wrap(*args):
+        out = f(*args)
+        # print(out)
+        return out
+    return wrap
+
+
+class ColorGenerator(object):
+    def __init__(self, dist=0.2, s_lb=0.6, v_lb=0.3, s=1.0, v=1.0):
+        self.counter = 0
+
+    def get_color(self):
+        out = BASE_COLORS[self.counter % len(BASE_COLORS)][self.counter // len(BASE_COLORS)]
+
+        self.counter = (self.counter + 1) % (len(BASE_COLORS) * len(BASE_COLORS[0]))
+
+        return hsv_to_rgb(*out)
+
+    def get_color_hex(self):
+        return rgb_to_hex(*self.get_color())
+
+class ValueRecorder(object):
+    def __init__(self, title, *args):
+        super(ValueRecorder, self).__init__()
+        self._color_generator = ColorGenerator(v=1.0, s_lb=0.75)
+        self.title    = title
+        group_colors  = [args[x] if type(args[x]) == tuple else (args[x], self._color_generator.get_color_hex()) for x in range(len(args))]
+        self.data     = {a: [] for a, _ in group_colors}
+        self.data_lim = {a: (1e20, -1e20) for a, _ in group_colors}
+        self.colors   = dict(group_colors)
+        self.thresholds = {}
+        self.x_labels = None
+        self.x_title  = None
+        self.x_space  = None
+        self.y_labels = None
+        self.y_title  = None
+        self.y_space  = None
+        self.grid     = False
+        self.legend_loc = None
+        self.marker   = None
+
+    def set_marker(self, marker):
+        self.marker = marker
+
+    def set_xspace(self, min, max):
+        self.x_space  = (min, max)
+
+    def set_xtitle(self, title):
+        self.x_title  = title
+
+    def set_xlabels(self, labels):
+        self.x_labels = labels
+
+    def set_yspace(self, min, max):
+        self.y_space  = (min, max)
+
+    def set_ytitle(self, title):
+        self.y_title  = title
+
+    def set_ylabels(self, labels):
+        self.y_labels = labels
+
+    def set_grid(self, grid):
+        self.grid = grid
+
+    def set_legend_location(self, loc):
+        self.legend_loc = loc
+
+    def log_data(self, group, value):
+        if group not in self.data:
+            raise Exception('Unknown data group "{}"'.format(group))
+        l, u = self.data_lim[group]
+        self.data_lim[group] = (min(value, l), max(value, u))
+        self.data[group].append(value)
+
+    def compute_limits(self):
+        for a, d in self.data.items():
+            self.data_lim[a] = (min(d), max(d))
+
+    def plot(self, ax):
+        if len(self.data_lim) == 0:
+            return
+
+        no_xticks = self.x_labels is not None and len(self.x_labels) == 0
+
+        if len(self.data) > 0:
+            labels = range(len(list(self.data.values())[0])) if self.x_labels is None or no_xticks else self.x_labels
+        else:
+            labels = []
+
+        if self.marker is None:
+            self.patches = [ax.plot(labels, d, color=self.colors[n], label=n)[0] for n, d in sorted(self.data.items())]
+        else:
+            self.patches = [ax.plot(labels, d, color=self.colors[n], label=n, marker=self.marker)[0] for n, d in sorted(self.data.items())]
+
+        data_lim_y = (float(min([l[0] for l in self.data_lim.values()])), float(max([l[1] for l in self.data_lim.values()])))
+
+        for n, (y, c) in self.thresholds.items():
+            ax.axhline(y, color=c)
+
+        if self.x_space is not None:
+            ax.set_xlim(self.x_space)
+
+        if self.y_space is not None:
+            ax.set_ylim(self.y_space)
+        else:
+            y_width = data_lim_y[1] - data_lim_y[0]
+            new_width = (data_lim_y[0] - y_width * 0.1, data_lim_y[1] + y_width * 0.1)
+            ax.set_ylim(new_width)
+
+        if self.y_labels is not None:
+            if type(self.y_labels) is tuple:
+                ax.set_yticks(self.y_labels[0])
+                ax.set_yticklabels(self.y_labels[1])
+            else:
+                ax.set_yticks(np.linspace(data_lim_y[0], data_lim_y[1], len(self.y_labels)))
+                ax.set_yticklabels(self.y_labels)
+
+        if self.x_title is not None:
+            ax.set_xlabel(self.x_title)
+
+        if self.y_title is not None:
+            ax.set_ylabel(self.y_title)
+
+        if no_xticks:
+            ax.set_xticklabels([])
+
+        loc = 'best' if self.legend_loc is None else self.legend_loc
+        ax.legend(handles=self.patches, loc=loc)
+        if self.title is not None:
+            ax.set_title(self.title)
+        ax.grid(self.grid)
+
+    def add_threshold(self, name, value, color=None):
+        if color is None:
+            color = BASE_COLORS[len(self.thresholds) % len(BASE_COLORS)]
+        self.thresholds[name] = (value, color)
+
+
+class SymbolicRecorder(ValueRecorder):
+    def __init__(self, title, **kwargs):
+        super(SymbolicRecorder, self).__init__(title, *kwargs.keys())
+        self.symbols = kwargs
+        self._labels, self._expressions = zip(*self.symbols.items()) if len(self.symbols) > 0 else ([], [])
+
+        temp_matrix = gm.Matrix(self._expressions)
+        self._expr_matrix = gm.speed_up(temp_matrix, gm.free_symbols(temp_matrix))
+
+    def log_symbols(self, subs_table):
+        np_matrix = self._expr_matrix(**{str(s): v for s, v in subs_table.items()})
+
+        for k, v in zip(self._labels, np_matrix.flatten()):
+            self.log_data(k, v)
